+++ test/SerdesNet.Tests/BinaryWriterTests.cs
@@ -171,7 +171,7 @@ namespace SerdesNet.Tests
         public void OffsetTests()
         {
             Write(x => Assert.Equal(SerializerFlags.Write, x.Flags));
-            Write(x => Assert.Equal(long.MaxValue, x.BytesRemaining));
+            Write(x => Assert.Equal(int.MaxValue, x.BytesRemaining));
             Write(x => Assert.False(x.IsComplete()));
 
             static void Block1(ISerializer s)
